chore: 폴더 구조 세팅 및 불필요한 모듈 삭제
