+++ src/domain/model/Lottos.js
@@ -1,6 +1,6 @@
-const { generateRandomLotto } = require('../../utils/RandomNumberGenerator');
+const { generateRandomLotto } = require('../../js/utils/RandomNumberGenerator');
 const Lotto = require('./Lotto');
-const { PROFIT_PER_RANK, PRICE_UNIT } = require('../../constants/constants');
+const { PROFIT_PER_RANK, PRICE_UNIT } = require('../../js/constants/constants');
 
 class Lottos {
   #lottos;
