+++ src/js/constants/constants.js
@@ -1,5 +1,3 @@
-const readline = require('readline');
-
 const CORRECT_COUNT_PER_RANK = Object.freeze({
   FIRST_RANK: 6,
   SECOND_RANK: 5,
@@ -63,11 +61,6 @@ const RESTART_COMMAND = Object.freeze({
   NO: 'n',
 });
 
-const RL = readline.createInterface({
-  input: process.stdin,
-  output: process.stdout,
-});
-
 const CONSOLE_MESSAGE = Object.freeze({
   ASK_PURCHASE_PRICE: '> 구입금액을 입력해 주세요.',
   showLottoCount: (lottoCount) => `${lottoCount}개를 구매했습니다.`,
@@ -103,7 +96,6 @@ module.exports = {
   PROFIT,
   PROFIT_PER_RANK,
   PRICE_UNIT,
-  RL,
   RANK,
   REGEX,
   RESTART_COMMAND,
